centralize clone and "parentTakesOwnership"-logic
This is necessary for the PDIList and PDISet (derived from PDIContainer) because their job is to update their item's "inUse" property.
They will update the property when adding or removing items if these conditions are met:
1. The object that contains the container takes ownership of the items
2. The object that contains the container is the "original" object - not a temporary clone.

The first condition is set when the list is created within its Data object and doesn't change at all later.

The second condition generally is the same. Because a container in a "original" Data object will always stay the original and the clone will always be the clone. Because of this, the "isClone" property is set when instanciating a parent item and its constructor will give the value to the data object.
But because it is possible to overwrite the entire data object with the =-operator, the setData() function ensures that all data is set but the isClone field doesn't change later.

In addition, when using the PDIContainer's =-operator, it will only apply it on its underlying Qt-based container class (QList or QHash) but will not affect the PDIContainer itself at all. So the parentOwnsItems and isInClone are not changeable in the context of the =-operator.