Busstop class: fix pointer issue with default platform
The problem was that the default platform was a pointer directly which pointed to the default platform. But when cloning the busstop (e.g. for editor dialogs), the platforms where cloned as well which lead do the problem that the defaultPlatform pointer pointed to a memory location that wasn't actually part of the platforms list.
I think, the easiest and best solution is, to just safe the default platforms ID instead and fetch the pointer dynamically every time it is requested.